pyproject.toml
[project]
name = "arbitrage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "black>=25.1.0",
    "celery>=5.5.3",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "django-ninja>=1.4.3",
    "flake8>=7.3.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "redis>=6.2.0",
]
